# dist/api.yml
openapi: 3.0.3
info:
  version: '1.0.0'
  title: 'Exercise 3.1'
  description: 'API - Web Development Environments'
...
paths:
...
  /users/favoritePlayers:
    get:
      tags:
        - Users
      summary: Return the favorite players that were saved by user
      responses:
        "200":
          description: data about the user's favorites players returned 
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPreview"       
        "401":
          description: User is not authorized
    post:
      summary: save player in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerID:
                  type: string
                  description: player's ID which the user saved
                  example: 82808
      responses:
        "201":
          description: The player successfully saved as favorite
        "303":
          description: The player already in favorites team  
        "400":
          description: One of the argument is not specified  
        "404":
          description: The player is not exists in superliga
  /users/favoriteTeams:
    get:
      tags:
        - Users
      summary: Return the favorite teams that were saved by user
      responses:
        "200":
          description: data about the user's favorites teams returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PreviewTeam"
        "401":
          description: User is not authorized

    post:
      summary: save team in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                teamID:
                  type: string
                  description: team's ID which the user saved
                  example: 85
      responses:
        "201":
          description: The team successfully saved as favorite
        "303":
          description: The team already in favorites teams  
        "400":
          description: One of the argument is not specified  
        "404":
          description: The team is not exists
  /users/favoriteGames:
    get:
      tags:
        - Users
      summary: Return the favorite games that were saved by user
      responses:
        "200":
          description: data about the user's favorites games returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GamePreview"
        "401":
          description: user is not authorized

    post:
      summary: save game in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                gameID:
                  type: string
                  description: game's ID which the user saved
                  example: 1
      responses:
        "201":
          description: The game successfully saved as favorite
        "303":
          description: The game already in favorites teams  
        "400":
          description: One of the argument is not specified  
        "401":
          description: The game already accured || The game is not exists 
...
components:
  schemas:
...
    PlayerPreview:
      type: object
      properties:
        playerID:
          type: number
          description: The player ID
          example: 4
        
        name:
          type: string
          description: The player name
          
        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png
...
    GamePreview:
      type: object
      properties:
        gameID:
          type: number
          description: The gameID
          example: 5

        game_date:
          type: string
          description: Game date
          example: '2021-06-01'

        game_time:
          type: string
          description: Game time
          example: '16:30'

        home_team_ID:
          type: number
          description: The home team ID
          example: 5
          
        away_team_ID:
          type: number
          description: The away team ID
          example: 15

        referee_ID:
          type: number
          description: The referee ID
          example: 3

        field:
          type: string 
          description: The field of the game        
...
    PreviewTeam:
      type: object
      properties:
        teamID:
          type: number
          description: The teamID
          example: 85

        name:
          type: string
          description: team name
          example: København

        players:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/PlayerPreview"  
        
        coach:
          type: string
          description: The team coach name
          example: "Alex Cohen"
...